[OPIK-6901] [BE] refactor: move the weekly-partition derivation into com.comet.opik.utils
The Monday-of-week derivation is a pure function of a UUID, and the same
partition expression backs both traces_local_v2 and spans_local_v2, so it
does not belong to TraceDAOImpl. Moved as-is to
com.comet.opik.utils.WeeklyPartitions, following SentinelTranslation: a
@UtilityClass whose javadoc carries the reasoning the callers must not
re-derive - above all that an empty result means "emit no predicate", not
"no partitions".

Pure move; no behaviour change. The test moves with it as
com.comet.opik.utils.WeeklyPartitionsTest.
