+++ apps/opik-backend/src/main/java/com/comet/opik/domain/TraceDAO.java
@@ -32,6 +32,7 @@ import com.comet.opik.utils.ClickHouseDateTimeFormat;
 import com.comet.opik.utils.ErrorUtils;
 import com.comet.opik.utils.JsonUtils;
 import com.comet.opik.utils.TruncationUtils;
+import com.comet.opik.utils.WeeklyPartitions;
 import com.comet.opik.utils.template.TemplateUtils;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
@@ -60,11 +61,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.math.BigDecimal;
-import java.time.DayOfWeek;
 import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneOffset;
-import java.time.temporal.TemporalAdjusters;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -1929,7 +1926,7 @@ class TraceDAOImpl implements TraceDAO {
      * <p>
      * {@code <if(partitions)>} adds the table's own weekly partition expression, bound as the exact set of partitions
      * the batch's ids resolve to. It is emitted <b>only when every id in the batch is a UUIDv7</b>
-     * ({@link #weeklyPartitionsOf}); if any id is not, the predicate is omitted and the statement is byte-identical to
+     * ({@link WeeklyPartitions#of}); if any id is not, the predicate is omitted and the statement is byte-identical to
      * the previous unbounded form. That preserves the original guarantee — a row whose {@code id_at} cannot be trusted
      * is still deleted, because no id in such a batch is used to derive a partition.
      * <p>
@@ -3527,41 +3524,6 @@ class TraceDAOImpl implements TraceDAO {
                 .doFinally(signalType -> endSegment(segment));
     }
 
-    /**
-     * The weekly partition values a batch of ids resolves to, or empty if the batch contains an id we refuse to derive
-     * a partition from.
-     * <p>
-     * Mirrors the table's partition expression exactly:
-     * {@code toYYYYMMDD(toDate32(id_at) - toIntervalDay(toDayOfWeek(id_at, 1)))}, where {@code id_at} is MATERIALIZED
-     * as {@code UUIDv7ToDateTime(toUUID(id))} — i.e. the Monday of the id's UTC week, as {@code yyyyMMdd}.
-     * <p>
-     * Returns empty unless <b>every</b> id is a UUIDv7. Deriving a partition from a non-v7 id would read whatever bits
-     * sit in the timestamp field, and a wrong partition means a <b>silently skipped delete</b>. All-or-nothing keeps the
-     * emitted SQL either fully pruned or exactly the previous unbounded form, never partially bounded.
-     * <p>
-     * Far-future ids are fine and deliberately supported: their {@code id_at} is bogus but self-consistent, so they live
-     * in the far-future partition this computes. Verified on prod-test — <b>0</b> partition mismatches across 11.23 M
-     * far-future rows.
-     */
-    static Optional<Set<Long>> weeklyPartitionsOf(Collection<UUID> ids) {
-        var partitions = new java.util.HashSet<Long>();
-
-        for (UUID id : ids) {
-            if (id == null || id.version() != 7) {
-                return Optional.empty();
-            }
-            // UUIDv7: the high 48 bits are the unix epoch in milliseconds.
-            long epochMilli = id.getMostSignificantBits() >>> 16;
-            LocalDate monday = Instant.ofEpochMilli(epochMilli)
-                    .atZone(ZoneOffset.UTC)
-                    .toLocalDate()
-                    .with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
-            partitions.add(monday.getYear() * 10000L + monday.getMonthValue() * 100L + monday.getDayOfMonth());
-        }
-
-        return partitions.isEmpty() ? Optional.empty() : Optional.of(partitions);
-    }
-
     @Override
     @WithSpan
     public Mono<Void> delete(Set<Pair<UUID, UUID>> projectIdTraceIdPairs, @NonNull Connection connection) {
@@ -3581,7 +3543,7 @@ class TraceDAOImpl implements TraceDAO {
                     var traceIds = batch.stream().map(pair -> pair.getRight().toString()).toArray(String[]::new);
 
                     // Prune to the batch's own partitions when every id is a UUIDv7; otherwise emit the unbounded form.
-                    var partitions = weeklyPartitionsOf(batch.stream().map(Pair::getRight).toList());
+                    var partitions = WeeklyPartitions.of(batch.stream().map(Pair::getRight).toList());
                     // Flag only, exactly like distributed_wrap: the values reach ClickHouse via the bind below,
                     // never through the template, so the rendered SQL is constant regardless of batch contents.
                     partitions.ifPresent(_ -> template.add("partitions", true));
